+++ lib/universal_internet_checker.dart
@@ -6,7 +6,16 @@ import 'package:http/http.dart' as http;
 class UniversalInternetChecker {
   static Uri doHService = DoHServices.cloudflare;
 
-  static Duration _interval = Duration(seconds: 30);
+  static final Duration _intervalOffline = Duration(seconds: 5);
+  static final Duration _intervalOnline = Duration(seconds: 25);
+  //nice if perfect multiples
+
+  //the number of iterations that the program should skip checking
+  static final int maxSkipCount =
+      _intervalOnline.inSeconds % _intervalOffline.inSeconds; //5 times
+
+  //variable to track runcount
+  late int currentSkipCount;
 
   //the domain to ask dns for
   static String checkAddress = 'google.com';
@@ -18,6 +27,7 @@ class UniversalInternetChecker {
   late Timer _timer;
 
   /// Stream controller to emit a notifier when connection status changes
+  /// TODO:static?
   final _streamController = StreamController<ConnectionStatus>.broadcast();
 
   /// Get method to return the stream from stream controller
@@ -42,20 +52,31 @@ class UniversalInternetChecker {
 
   void _setupPolling() {
     _checkAndBroadcast();
-    _timer = Timer.periodic(_interval, _checkAndBroadcast);
+    _timer = Timer.periodic(_intervalOffline, _checkAndBroadcast);
   }
 
   /// Method to check the connection status according the duration
   void _checkAndBroadcast([Timer? timer]) async {
     if (!_streamController.hasListener) return;
 
-    ConnectionStatus currentStatus = await checkInternet();
+    ConnectionStatus _currentStatus = await checkInternet();
 
-    if (_lastStatus != currentStatus && _streamController.hasListener) {
-      _streamController.add(currentStatus);
-      print('connection status changed: $currentStatus');
-      _lastStatus = currentStatus;
+    if (_currentStatus == _lastStatus) return; //do nothing if nothing to update
+    //if last status is online, check only if we have not checked in 10 seconds
+    if (_lastStatus == ConnectionStatus.online &&
+        _currentStatus == ConnectionStatus.offline) {
+      //if it comes here, then at least one bradcast has happened
+      //because initial status is ConnectionStatus.unknown
+      if (currentSkipCount < maxSkipCount) {
+        //don't update listener for 1st, 2nd, 3rd and 4th run if online
+        currentSkipCount += 1;
+        return;
+      } //broadcast offline status if 5th count
     }
+    _streamController.add(_currentStatus);
+    print('connection status changed: $_currentStatus');
+    _lastStatus = _currentStatus;
+    currentSkipCount = 1; //reset the counter if broadcast
   }
 
   /// Static method to check if it's connected to internet
@@ -71,7 +92,7 @@ class UniversalInternetChecker {
           }),
           headers: {
             'Accept': 'application/dns-json'
-          }).timeout(Duration(milliseconds: 1200));
+          }).timeout(Duration(milliseconds: 4500));
 
       // Close & retrive response
       if (response.statusCode == 200) {