+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## 0.1.2
 
 - More sensible 30 second check interval (debounce)
+
+## 0.1.3
+
+- Due to debounce issues, listeners will only be updated 
+- every 5 seconds if offline
+- every 25 seconds if online